move chore and scheduling classes to their own files
