+++ lib/chores.rb
@@ -1,0 +1,3 @@
+require_relative './schedulings'
+
+Dir.glob("#{File.dirname(__FILE__)}/chores/*.rb").each(&method(:require))
