+++ lib/schedulings.rb
@@ -1,0 +1,1 @@
+Dir.glob("#{File.dirname(__FILE__)}/schedulings/*.rb").each(&method(:require))
