+++ lib/schedulings/daily_scheduling.rb
@@ -1,0 +1,7 @@
+require_relative './weekly_scheduling'
+
+class DailyScheduling < WeeklyScheduling
+  def initialize
+    @days = [:monday, :tuesday, :wednesday, :thursday, :friday]
+  end
+end
